all good

+++ extras/get_time.c
@@ -10,12 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "philos.h"
+#include "extras.h"
 
-int get_time(void)
+size_t get_time(void)
 {
     struct timeval time;
     
     gettimeofday(&time, NULL);
-    return (time.tv_sec * 1000 + time.tv_usec / 1000);
+    return (size_t)(time.tv_sec * 1000 + (time.tv_usec / 1000));
 }
